perf: Extract ChatInputForm from ChatPanel to avoid re-renders
Extracted the input state and form logic from `ChatPanel.tsx` into a separate `ChatInputForm.tsx` component. This prevents the entire `ChatPanel` (including all historical message bubbles and markdown parsing) from re-rendering on every keystroke, which was causing O(N) re-renders.
