+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/api/tauri.ts","./src/components/AddCharacterModal.tsx","./src/components/CharacterSelect.tsx","./src/components/ChatPanel.tsx","./src/components/Live2DCanvas.tsx","./src/components/LoadingOverlay.tsx","./src/components/MemoryStatePanel.tsx","./src/components/MicButton.tsx","./src/components/MiniFloatingMessages.tsx","./src/components/MiniToolPill.tsx","./src/components/MiniWidget.tsx","./src/components/ModelSettings.tsx","./src/components/Onboarding.tsx","./src/components/Settings.tsx","./src/components/ToolCallBubble.tsx","./src/components/VRMCanvas.tsx","./src/constants/bgPresets.ts","./src/hooks/useAudioAnalyser.ts","./src/hooks/useAudioQueue.ts","./src/hooks/useChat.ts","./src/hooks/useLive2D.ts","./src/hooks/useVRM.ts","./src/hooks/useVoice.ts","./src/hooks/useVoiceConfirmation.ts","./src/hooks/useWindow.ts","./src/types/index.ts","./src/utils/mixamoRigMap.ts"],"version":"5.9.3"}
+{"root":["./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/api/tauri.test.ts","./src/api/tauri.ts","./src/components/AddCharacterModal.tsx","./src/components/CharacterSelect.tsx","./src/components/ChatInputForm.tsx","./src/components/ChatPanel.tsx","./src/components/Live2DCanvas.tsx","./src/components/LoadingOverlay.test.tsx","./src/components/LoadingOverlay.tsx","./src/components/MemoryStatePanel.tsx","./src/components/MicButton.tsx","./src/components/MiniFloatingMessages.tsx","./src/components/MiniToolPill.tsx","./src/components/MiniWidget.tsx","./src/components/ModelSettings.tsx","./src/components/Onboarding.tsx","./src/components/Settings.tsx","./src/components/ToolCallBubble.tsx","./src/components/VRMCanvas.tsx","./src/constants/bgPresets.ts","./src/hooks/useAudioAnalyser.ts","./src/hooks/useAudioQueue.ts","./src/hooks/useChat.ts","./src/hooks/useLive2D.ts","./src/hooks/useVRM.ts","./src/hooks/useVoice.ts","./src/hooks/useVoiceConfirmation.ts","./src/hooks/useWindow.ts","./src/test/setup.ts","./src/types/index.ts","./src/utils/mixamoRigMap.ts"],"version":"5.9.3"}
